# openapi.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: A simple REST API for managing tasks
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /tasks:
    get:
      summary: List all tasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks/{task_id}:
    get:
      summary: Get a specific task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task deleted successfully
        '400':
          description: Invalid task ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          minLength: 1
          example: "Write comprehensive API documentation"
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-12-01T12:00:00Z"

    TaskCreate:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          minLength: 1
          example: "Write comprehensive API documentation"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          minLength: 1
          example: "Write comprehensive API documentation"
        completed:
          type: boolean
          example: false

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Task not found"