# src/api/main.py
from fastapi import FastAPI, HTTPException, Request
from fastapi.exceptions import RequestValidationError
from fastapi.responses import JSONResponse, Response
from starlette.exceptions import HTTPException as StarletteHTTPException

from .database import db
from .models import Task, TaskCreate, TaskUpdate

app = FastAPI(
    title="Task Management API",
    description="A simple REST API for managing tasks",
    version="1.0.0",
    redirect_slashes=False,
)


@app.exception_handler(RequestValidationError)
async def validation_exception_handler(
    request: Request, exc: RequestValidationError
) -> JSONResponse:
    return JSONResponse(
        status_code=400,
        content={"message": "Invalid input"},
    )


@app.exception_handler(StarletteHTTPException)
async def custom_http_exception_handler(
    request: Request, exc: StarletteHTTPException
) -> JSONResponse:
    headers = {}
    if exc.status_code == 405:
        # Add Allow header for Method Not Allowed responses
        if request.url.path == "/tasks":
            headers["Allow"] = "GET, POST"
        elif (
            request.url.path.startswith("/tasks/") and request.url.path.count("/") == 2
        ):
            headers["Allow"] = "GET, PUT, DELETE"

    return JSONResponse(
        status_code=exc.status_code,
        content={"message": exc.detail},
        headers=headers,
    )


@app.get("/tasks", response_model=list[Task])
def get_tasks() -> list[Task]:
    return db.get_all_tasks()


@app.post("/tasks", response_model=Task, status_code=201)
def create_task(task: TaskCreate) -> Task:
    return db.create_task(task)


@app.get("/tasks/{task_id}", response_model=Task)
def get_task(task_id: int) -> Task:
    task = db.get_task(task_id)
    if task is None:
        raise HTTPException(status_code=404, detail="Task not found")
    return task


@app.put("/tasks/{task_id}", response_model=Task)
def update_task(task_id: int, task_update: TaskUpdate) -> Task:
    task = db.update_task(task_id, task_update)
    if task is None:
        raise HTTPException(status_code=404, detail="Task not found")
    return task


@app.delete("/tasks/{task_id}", status_code=204)
def delete_task(task_id: int) -> Response:
    success = db.delete_task(task_id)
    if not success:
        raise HTTPException(status_code=404, detail="Task not found")
    return Response(status_code=204)